Add files via upload
The goal of this project is to develop a deep learning model capable of summarizing long documents or articles into concise and coherent summaries. Text summarization is a crucial task in natural language processing (NLP) that helps in quickly extracting the most important information from large texts. This project uses a sequence-to-sequence model with an attention mechanism, specifically utilizing the Transformer architecture, which has shown state-of-the-art performance in various NLP tasks.